refactor: ReplyServiceImpl 삭제&병합 리팩토링
